Moved some stuff around and added a cmake
