+++ Source/Surface.cpp
@@ -22,9 +22,9 @@ void RenderTarget::SetPixelColour(int x, int y, Colour pixelColour)
 {
 	double maxChannel = 255;
 	// guarantee pixelColour is between 0 and 1.
-	_rBuffer[x*_Width + y] = unsigned char(pixelColour[0] * maxChannel);
-	_gBuffer[x*_Width + y] = unsigned char(pixelColour[1] * maxChannel);
-	_bBuffer[x*_Width + y] = unsigned char(pixelColour[2] * maxChannel);
+	_rBuffer[x*_Width + y] = static_cast<unsigned char>(pixelColour[0] * maxChannel);
+	_gBuffer[x*_Width + y] = static_cast<unsigned char>(pixelColour[1] * maxChannel);
+	_bBuffer[x*_Width + y] = static_cast<unsigned char>(pixelColour[2] * maxChannel);
 }
 
 unsigned char* RenderTarget::GetRedChannel()
