--- a3/raytracerLinux/.directory
@@ -1,3 +1,0 @@
-[Dolphin]
-PreviewsShown=true
-Timestamp=2015,12,6,21,40,4
