--- a3/raytracerW32/raytracer.sln
@@ -1,27 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30501.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "raytracer", "raytracer.vcxproj", "{38CA0D5B-AB0B-46BF-B70D-4D32E622D2B1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Debug|x64 = Debug|x64
-		Release|Win32 = Release|Win32
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{38CA0D5B-AB0B-46BF-B70D-4D32E622D2B1}.Debug|Win32.ActiveCfg = Debug|Win32
-		{38CA0D5B-AB0B-46BF-B70D-4D32E622D2B1}.Debug|Win32.Build.0 = Debug|Win32
-		{38CA0D5B-AB0B-46BF-B70D-4D32E622D2B1}.Debug|x64.ActiveCfg = Debug|x64
-		{38CA0D5B-AB0B-46BF-B70D-4D32E622D2B1}.Debug|x64.Build.0 = Debug|x64
-		{38CA0D5B-AB0B-46BF-B70D-4D32E622D2B1}.Release|Win32.ActiveCfg = Release|Win32
-		{38CA0D5B-AB0B-46BF-B70D-4D32E622D2B1}.Release|Win32.Build.0 = Release|Win32
-		{38CA0D5B-AB0B-46BF-B70D-4D32E622D2B1}.Release|x64.ActiveCfg = Release|x64
-		{38CA0D5B-AB0B-46BF-B70D-4D32E622D2B1}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
