+++ api/Euclidean.h
