+++ api/Intersection.h
